Initial dummy cache implementation.
Have a nss function as the first item, so it's always hit first.
on_status is always ACT_CONTINUE, since the cache should be transparent.
There's a chance this would create issues with "not found" results; we
could consider a separate set of implementations for "found" and "not
found".

For now, caching is opt-in via the -C command line arg.