Start integrating caching into socket_handle.
In return_result, change the outer for loop into a do-while one, so we
can first check the cache results and later the actual modules. This
requires adding the cache_run variable to track if we are using the
cache module or the configured modules.

We also move from using sysconf every loop to find out buffer size, to
using sizes initialized in init_socket_handling.

Also move many variable declarations into the do-while loop, since they
don't need to be available for the rest of the function; this commit
temporarily causes a memory leak in with buf, which will be fixed in the
following commits.