Simplify fallback write code.
Testing if l is 0 is unnecessary: since this always happens after the
loop which necessarily exits when l is set to 0, and which in any other
situation will have already returned (thanks to fixes in previous
commits), it should always run if reached. Use the same style of
conditionals as other parts of the code instead of cases.