Fix cache modules initialization.
We were using the wrong actions, ACT_RETURN should be only for SUCCESS,
not any other. All others should be ACT_CONTINUE.

Initialize on_status for the modules in the module definitions. This
fixes a bug where on_status would be full of ACT_RETURN (numerical value
of 0) when cache was disabled, because initializing the arrays only
happened in init_caches.