Fix cache run detection and memory management in return_result.
Since cache_run is set to false early on, we need a separate variable to
track if we are inside a cache run. That variable is using_cache.  Using
using_cache instead of cache_run fixes a bug in the cache addition
block, where it would always try to add entries to the cache, even when
the entry itself came from the cache - this would result in unnecessary
write-locking of the cache, just to find there was a valid entry in the
cache already.

We also add a FREE_ALLOC macro to simplify dealing with buf/res.l.grps;
this way we need only to call the macro and it will perform the correct
free() call(s). Thanks to the macro, we can easily make res.l.groups
still valid after write_groups has used it, and guarantee that we always
free it. We also a fix memory leak for when the action for status isn't
RETURN.