Implement LRU-like part of cache.
When not overwriting an outdated entry for the query itself, we should
overwrite the oldest invalid query available.

For consistent checking and possible efficiency, we store the result of
monotonic_seconds() and use it in compare_timestamps() instead of using
validate_timestamp().

We use oldest and oldest_i to track, respectively, the lowest t value in
the cache and the index of said value.  Initializing oldest is tricky,
because time_t doesn't have an easily accessible max value, so we use
the first entry in the cache, but that can't be accessed if the cache is
empty, so we also check for that.

If we have found an invalid entry that should be overwritten, we free
its storage that can't be reused (for passwd and group that's the char
buffer; for initgroups that's the grps array and name). The code that
increments cache length and requires multiple allocations is then moved
to the conditional that runs when no invalid entry has been found. This
code allocates the new structs it has to. For the case with passwd and
groups, this also fixes the inconsistent case where allocating a struct
passwd or group failed.