Fix race-condition in accessing cache storage.
The initial design was implemented based on the idea that access to
the cache's underlying storage would avoid allocations and generally
speaking lead to a cleaner design. Unfortunately, this also opened up
the program to use-after-free issues.

A sequence of events as explained below shows a possible scenario (Tx
are threads):

1. T1 enters cache, finds valid entry, return it (returns buffer from
   cache)
2. T2 enters cache, finds the entry is invalid now, doesn't return it;
   finds new entry from nss module
3. T2 tries to add new entry to cache, will release underlying memory
   (free(buffer))
4. T1 finally gets to writing out, for _whatever reason_, and now buffer
   is UAF

There are two possible ways to deal with this:

- add a mutex to struct, lock while it's being used; means that the
  wrlock will be held for a potentially very *long* time, which means
  holding up the rdlock for consumers. Therefore, not an option
- actually allocate memory when reading from cache; not ideal but
  actually avoids all of the lifetime issues

We went with the latter, but kept most of return_result intact, which
means there are some gnarly pointer conversions going on. The code
should be restructured, instead, to call the cache query functions
separately from the other modules.