Implement a pre-threaded server model.
Instead of launching a thread to listen to the next request, have a
given number (for now, 5) of server threads pre-launched and a listener
thread that sends queries to the server threads. This way, we improve
latency considerably when more than one client is making queries at the
same time.

serv_thread() is the logic after the accept() call in socket_handle(),
plus the semaphore handling. socket_handle() only calls accept and sends
requests to the server threads. Since socket_handle() is only called
once now, we can remove most of its parameters; initialization of
locale_t l is moved to init_socket_handling.